Refactor parser/piipeline logic to simplify
